+++ packages/playground/.env
@@ -27,4 +27,4 @@ VITE_ENABLE_TOOLS=true
 VITE_ENABLE_KNOWLEDGE=false
 VITE_ENABLE_NEW_KNOWLEDGE=false
 VITE_ENABLE_ATTACHMENT=true
-VITE_ENABLE_DISCOVER=false
+VITE_ENABLE_AGENT=false
